client: C++ std::min/max fixes, remove reading from unused data from mstudioseqgroup

+++ cl_dll/entity.cpp
@@ -9,6 +9,7 @@
 
 #include <memory.h>
 #include <math.h>
+#include <cstdint>
 
 #include "hud.h"
 #include "cl_util.h"
@@ -542,8 +543,8 @@ void DLLEXPORT HUD_TempEntUpdate(
 				s = sin( pTemp->entity.baseline.origin[2] + fastFreq );
 				c = cos( pTemp->entity.baseline.origin[2] + fastFreq );
 
-				pTemp->entity.origin[0] += pTemp->entity.baseline.origin[0] * frametime + 8 * sin( client_time * 20 + (int)pTemp );
-				pTemp->entity.origin[1] += pTemp->entity.baseline.origin[1] * frametime + 4 * sin( client_time * 30 + (int)pTemp );
+				pTemp->entity.origin[0] += pTemp->entity.baseline.origin[0] * frametime + 8 * sin( client_time * 20 + (intptr_t)pTemp );
+				pTemp->entity.origin[1] += pTemp->entity.baseline.origin[1] * frametime + 4 * sin( client_time * 30 + (intptr_t)pTemp );
 				pTemp->entity.origin[2] += pTemp->entity.baseline.origin[2] * frametime;
 			}
